Small fix
git-svn-id: https://svn.viathinksoft.com/svn/alarming/trunk@10 58d948ce-2912-478e-bbcd-812bb8b1fcef

+++ Server/daemon.py
@@ -198,7 +198,8 @@ function onload() {
 		if config.enable_motion_detect:
 			os.system(os.path.dirname(os.path.abspath(__file__)) + "/motion/motion_stop_safe")
 		if config.enable_doorbell_listener:
-	                sp.Popen.terminate(g_bellListenerProc)
+			if g_bellListenerProc != None:
+		                sp.Popen.terminate(g_bellListenerProc)
 
 	self._output(200, '')
